requirements.txt
fastapi
uvicorn
psycopg2-binary
passlib
onnxruntime
onnx
transformers
torch
python-dotenv
pytest
httpx
